Fixed typo in basic_hub_link.png (#49)
* Fixed typo in basic_hub_link.png

Edited hub table name from HUB_A to HUB_B and replaced .png and .drawio file for the Effectivity Satellites document

* Fixed colour of arrows in basic_hub_link.png

Made the arrows white with a black outline so they are visible in both light and dark mode. This is present in both .png and .drawio files.

* Fixed colour of arrows in basic_hub_link.png

AB#2262 Made the arrows white with a black outline so they are visible in both light and dark mode. This is present in both .png and .drawio files.